Create tree to plot

Create phylogenetic tree with edge lengths according to a shape

phyloshape(
  shape,
  tree = NULL,
  point = NULL,
  nlines = 360,
  method = "fill",
  depth.k = 0.95
)

Arguments

shape

a SpatialVector object with only one geometry whose contour will be used to shape the phylogeny

tree

Phylogenetic tree to midify its branch lengths. If NULL, a tree of random topology will be created. The number of tips will be randomly pruned to match 'nlines'

point

a SpatialVector point from where the phylogeny will be plotted. If NULL, the function will use the centroid.

nlines

number of lines (tips) the resulting phylogeny will include. default is 360.

method

Method used to calculate the branch lengths. Method 'extend' will mantain the original phylogeny shape and add distances just to the tip branches. Method 'fill' will calculate distances from the tips back to the root. Distances from a tip or node to its direct ancestor are calculated using a depth constant.

depth.k

Applies for method 'fill'. Percentage of branch that separates internal nodes. Values near to 0 have clustered internal nodes near to the root, thus resemble to a polytomy, while values near 1 have internal nodes clustered toward tips. Default value is 0.95.

Value

A phylogenetic tree whose edge lengths will draw the specified shape.

Author

Ignacio Ramos-Gutiérrez

Examples

if (FALSE) { # interactive()
 tree2 <- phyloshape(tree=NULL, shape,  method = "fill", nlines = 150, depth.k = 0.99)
 ape::plot.phylo(tree2, type = "f", show.tip.label = F)
 ape::tiplabels(pch=16, cex=.5)
}